src/main/java/org/acme/adapters/out/InMemoryTaskRepository.java
package org.acme.adapters.out;

import org.acme.dominio.Task;
import org.acme.dominio.TaskRepository;

import jakarta.enterprise.context.ApplicationScoped;

import java.util.ArrayList;
import java.util.List;
import java.util.Optional;

@ApplicationScoped
public class InMemoryTaskRepository implements TaskRepository {
    private final List<Task> tasks = new ArrayList<>();

    @Override
    public List<Task> findAll() {
        return new ArrayList<>(tasks);
    }

    @Override
    public Optional<Task> findById(String id) {
        return tasks.stream().filter(task -> task.getId().equals(id)).findFirst();
    }

    @Override
    public void save(Task task) {
        tasks.removeIf(t -> t.getId().equals(task.getId()));
        tasks.add(task);
    }

    @Override
    public void deleteById(String id) {
        tasks.removeIf(task -> task.getId().equals(id));
    }
}
